+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All notable changes to this project will be documented in this file.
 - Field `read_header` added to the `amqp_1` input.
 - A debug endpoint `/debug/pprof/allocs` has been added for profiling allocations.
 - New `cockroachdb_changefeed` input.
+- The `open_telemetry_collector` tracer now supports sampling.
 
 ### Fixed